move rest of odbc/api definitions to new crate
