+++ definitions/src/bulk_operation.rs
@@ -1,9 +1,11 @@
-/// Supported `BulkOperation` operations
-#[repr(u16)]
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+use num_derive::FromPrimitive;
+
+#[allow(non_camel_case_types)]
+#[repr(i16)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy, FromPrimitive)]
 pub enum BulkOperation {
-    Add = 4,
-    UpdateByBookmark = 5,
-    DeleteByBookmark = 6,
-    FetchByBookmark = 7,
+    SQL_ADD = 4,
+    SQL_UPDATE_BY_BOOKMARK = 5,
+    SQL_DELETE_BY_BOOKMARK = 6,
+    SQL_FETCH_BY_BOOKMARK = 7,
 }
