+++ definitions/src/fetch_orientation.rs
@@ -1,15 +1,15 @@
-/// Codes used for FetchOrientation in `SQLFetchScroll`, `SQLDataSources` and in `SQLDrivers`
+use num_derive::FromPrimitive;
+
+#[allow(non_camel_case_types)]
 #[repr(u16)]
-#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+#[derive(Debug, PartialEq, Eq, Clone, Copy, FromPrimitive)]
 pub enum FetchOrientation {
-    Next = 1,
-    First = 2,
-    // Other codes used for FetchOrientation in SQLFetchScroll()
-    Last = 3,
-    Prior = 4,
-    Absolute = 5,
-    Relative = 6,
-    // additional SQLDataSources fetch directions
-    FirstUser = 31,
-    FirstSystem = 32,
+    SQL_FETCH_NEXT = 1,
+    SQL_FETCH_FIRST = 2,
+    SQL_FETCH_LAST = 3,
+    SQL_FETCH_PRIOR = 4,
+    SQL_FETCH_ABSOLUTE = 5,
+    SQL_FETCH_RELATIVE = 6,
+    SQL_FETCH_FIRST_USER = 31,
+    SQL_FETCH_FIRST_SYSTEM = 32,
 }
