+++ integration_test/src/test_runner/test_generator_util.rs
@@ -129,46 +129,46 @@ pub fn generate_baseline_test_file(entry: &TestEntry, stmt: HStmt) -> Result<()>
 fn get_expected_data_type(sql_type: &Value) -> CDataType {
     match sql_type {
         Value::Number(n) => match FromPrimitive::from_i64(n.as_i64().unwrap()) {
-            Some(SqlDataType::SQL_UNKNOWN_TYPE) => CDataType::Char,
-            Some(SqlDataType::SQL_CHAR) => CDataType::Char,
-            Some(SqlDataType::SQL_NUMERIC) => CDataType::Numeric,
-            Some(SqlDataType::SQL_DECIMAL) => CDataType::Numeric,
-            Some(SqlDataType::SQL_INTEGER) => CDataType::SLong,
-            Some(SqlDataType::SQL_SMALLINT) => CDataType::SShort,
-            Some(SqlDataType::SQL_FLOAT) => CDataType::Float,
-            Some(SqlDataType::SQL_REAL) => CDataType::Numeric,
-            Some(SqlDataType::SQL_DOUBLE) => CDataType::Double,
-            Some(SqlDataType::SQL_DATETIME) => CDataType::TypeTimestamp,
-            Some(SqlDataType::SQL_VARCHAR) => CDataType::Char,
-            Some(SqlDataType::SQL_TYPE_DATE) => CDataType::TypeDate,
-            Some(SqlDataType::SQL_TYPE_TIME) => CDataType::TypeTime,
-            Some(SqlDataType::SQL_TYPE_TIMESTAMP) => CDataType::Char,
-            Some(SqlDataType::SQL_TIME_OR_INTERVAL) => CDataType::Char,
-            Some(SqlDataType::SQL_TIMESTAMP) => CDataType::Default,
-            Some(SqlDataType::SQL_LONGVARCHAR) => CDataType::Char,
-            Some(SqlDataType::SQL_BINARY) => CDataType::Binary,
-            Some(SqlDataType::SQL_VARBINARY) => CDataType::Binary,
-            Some(SqlDataType::SQL_LONGVARBINARY) => CDataType::Binary,
-            Some(SqlDataType::SQL_BIGINT) => CDataType::SBigInt,
-            Some(SqlDataType::SQL_TINYINT) => CDataType::STinyInt,
-            Some(SqlDataType::SQL_BIT) => CDataType::Bit,
-            Some(SqlDataType::SQL_WCHAR) => CDataType::WChar,
-            Some(SqlDataType::SQL_WVARCHAR) => CDataType::WChar,
-            Some(SqlDataType::SQL_WLONGVARCHAR) => CDataType::WChar,
-            Some(SqlDataType::SQL_GUID) => CDataType::Guid,
-            Some(SqlDataType::SQL_INTERVAL_DAY) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_YEAR) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_MONTH) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_HOUR) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_MINUTE) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_SECOND) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_YEAR_TO_MONTH) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_DAY_TO_HOUR) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_DAY_TO_MINUTE) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_DAY_TO_SECOND) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_HOUR_TO_MINUTE) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_HOUR_TO_SECOND) => CDataType::Char,
-            Some(SqlDataType::SQL_INTERVAL_MINUTE_TO_SECOND) => CDataType::Char,
+            Some(SqlDataType::SQL_UNKNOWN_TYPE) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_CHAR) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_NUMERIC) => CDataType::SQL_C_NUMERIC,
+            Some(SqlDataType::SQL_DECIMAL) => CDataType::SQL_C_NUMERIC,
+            Some(SqlDataType::SQL_INTEGER) => CDataType::SQL_C_SLONG,
+            Some(SqlDataType::SQL_SMALLINT) => CDataType::SQL_C_SSHORT,
+            Some(SqlDataType::SQL_FLOAT) => CDataType::SQL_C_FLOAT,
+            Some(SqlDataType::SQL_REAL) => CDataType::SQL_C_NUMERIC,
+            Some(SqlDataType::SQL_DOUBLE) => CDataType::SQL_C_DOUBLE,
+            Some(SqlDataType::SQL_DATETIME) => CDataType::SQL_C_TYPE_TIMESTAMP,
+            Some(SqlDataType::SQL_VARCHAR) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_TYPE_DATE) => CDataType::SQL_C_TYPE_DATE,
+            Some(SqlDataType::SQL_TYPE_TIME) => CDataType::SQL_C_TYPE_TIME,
+            Some(SqlDataType::SQL_TYPE_TIMESTAMP) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_TIME_OR_INTERVAL) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_TIMESTAMP) => CDataType::SQL_C_DEFAULT,
+            Some(SqlDataType::SQL_LONGVARCHAR) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_BINARY) => CDataType::SQL_C_BINARY,
+            Some(SqlDataType::SQL_VARBINARY) => CDataType::SQL_C_BINARY,
+            Some(SqlDataType::SQL_LONGVARBINARY) => CDataType::SQL_C_BINARY,
+            Some(SqlDataType::SQL_BIGINT) => CDataType::SQL_C_SBIGINT,
+            Some(SqlDataType::SQL_TINYINT) => CDataType::SQL_C_STINYINT,
+            Some(SqlDataType::SQL_BIT) => CDataType::SQL_C_BIT,
+            Some(SqlDataType::SQL_WCHAR) => CDataType::SQL_C_WCHAR,
+            Some(SqlDataType::SQL_WVARCHAR) => CDataType::SQL_C_WCHAR,
+            Some(SqlDataType::SQL_WLONGVARCHAR) => CDataType::SQL_C_WCHAR,
+            Some(SqlDataType::SQL_GUID) => CDataType::SQL_C_GUID,
+            Some(SqlDataType::SQL_INTERVAL_DAY) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_YEAR) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_MONTH) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_HOUR) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_MINUTE) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_SECOND) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_YEAR_TO_MONTH) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_DAY_TO_HOUR) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_DAY_TO_MINUTE) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_DAY_TO_SECOND) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_HOUR_TO_MINUTE) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_HOUR_TO_SECOND) => CDataType::SQL_C_CHAR,
+            Some(SqlDataType::SQL_INTERVAL_MINUTE_TO_SECOND) => CDataType::SQL_C_CHAR,
             None => unreachable!("invalid sql_type encountered: {:?}", sql_type),
         },
         v => unreachable!("sql_type should always be a number: {:?}", v),
